src/test/kotlin/dev/shtanko/algorithms/leetcode/SequentialDigitsTest.kt
/*
 * Copyright 2024 Oleksii Shtanko
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package dev.shtanko.algorithms.leetcode

import java.util.stream.Stream
import org.assertj.core.api.Assertions
import org.junit.jupiter.api.extension.ExtensionContext
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.Arguments
import org.junit.jupiter.params.provider.ArgumentsProvider
import org.junit.jupiter.params.provider.ArgumentsSource

abstract class SequentialDigitsTest<out T : SequentialDigits>(private val strategy: T) {
    private class InputArgumentsProvider : ArgumentsProvider {
        override fun provideArguments(context: ExtensionContext?): Stream<out Arguments> = Stream.of(
            Arguments.of(
                100,
                300,
                listOf(123, 234),
            ),
            Arguments.of(
                1000,
                13000,
                listOf(1234, 2345, 3456, 4567, 5678, 6789, 12345),
            ),
        )
    }

    @ParameterizedTest
    @ArgumentsSource(InputArgumentsProvider::class)
    fun `sequential digits test`(low: Int, high: Int, expected: List<Int>) {
        val actual = strategy(low, high)
        Assertions.assertThat(actual).isEqualTo(expected)
    }
}

class SequentialDigitsIterativeTest : SequentialDigitsTest<SequentialDigits>(SequentialDigitsIterative())
